+++ ztu/lib/screens/home/home_screen.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:ztu/models/category_model.dart';
 import 'package:ztu/widgets/custom_appbar.dart';
 import 'package:ztu/widgets/custom_nav_bar.dart';
+import 'package:ztu/widgets/hero_carousel_card.dart';
 
 class HomeScreen extends StatelessWidget {
   static const String routeName = '/';
@@ -28,56 +29,8 @@ class HomeScreen extends StatelessWidget {
               enlargeCenterPage: true,
               enlargeStrategy: CenterPageEnlargeStrategy.height),
           items: Category.categories
-              .map((categoryItem) => HeroCarouseCard(category: categoryItem))
+              .map((categoryItem) => HeroCarouselCard(category: categoryItem))
               .toList(),
         ));
-  }
-}
-
-class HeroCarouseCard extends StatelessWidget {
-  final Category category;
-
-  const HeroCarouseCard({super.key, required this.category});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: const EdgeInsets.all(5.0),
-      child: ClipRRect(
-        borderRadius: const BorderRadius.all(Radius.circular(5.0)),
-        child: Stack(
-          children: <Widget>[
-            Image.network(category.imgUrl, fit: BoxFit.cover, width: 1000.0),
-            Positioned(
-              bottom: 0.0,
-              left: 0.0,
-              right: 0.0,
-              child: Container(
-                decoration: const BoxDecoration(
-                  gradient: LinearGradient(
-                    colors: [
-                      Color.fromARGB(200, 0, 0, 0),
-                      Color.fromARGB(0, 0, 0, 0)
-                    ],
-                    begin: Alignment.bottomCenter,
-                    end: Alignment.topCenter,
-                  ),
-                ),
-                padding: const EdgeInsets.symmetric(
-                    vertical: 10.0, horizontal: 20.0),
-                child: Text(
-                  category.name,
-                  style: const TextStyle(
-                    color: Colors.white,
-                    fontSize: 20.0,
-                    fontWeight: FontWeight.bold,
-                  ),
-                ),
-              ),
-            ),
-          ],
-        ),
-      ),
-    );
   }
 }
